+++ client/CHANGELOG.md
@@ -9,3 +9,16 @@ All notable changes to this project will be documented in this file.
 -   Soical Authentication
 -   Create Swatches, Palettes, Gradients
 -   Like / Unlike Colors
+
+---
+
+## [1.1.0] - 2019-01-16
+
+### Added
+
+-   Tooltip on Card hover
+-   Copy to clipboard option
+
+### Removed
+
+-   Container Component